+++ apps/client/src/components/dashboard/useTaskCreationForm.tsx
@@ -1,0 +1,843 @@
+import type { EditorApi } from "@/components/dashboard/DashboardInput";
+import { GitHubIcon } from "@/components/icons/github";
+import { useTheme } from "@/components/theme/use-theme";
+import type { SelectOption } from "@/components/ui/searchable-select";
+import {
+  Tooltip,
+  TooltipContent,
+  TooltipTrigger,
+} from "@/components/ui/tooltip";
+import { useExpandTasks } from "@/contexts/expand-tasks/ExpandTasksContext";
+import { useSocket } from "@/contexts/socket/use-socket";
+import { createFakeConvexId } from "@/lib/fakeConvexId";
+import { attachTaskLifecycleListeners } from "@/lib/socket/taskLifecycleListeners";
+import { branchesQueryOptions } from "@/queries/branches";
+import { api } from "@cmux/convex/api";
+import type { Doc, Id } from "@cmux/convex/dataModel";
+import type {
+  ProviderStatusResponse,
+  TaskAcknowledged,
+  TaskError,
+  TaskStarted,
+} from "@cmux/shared";
+import { AGENT_CONFIGS } from "@cmux/shared/agentConfig";
+import { convexQuery } from "@convex-dev/react-query";
+import { useQuery } from "@tanstack/react-query";
+import { useAction, useMutation } from "convex/react";
+import { Server as ServerIcon } from "lucide-react";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { toast } from "sonner";
+import { z } from "zod";
+
+const DEFAULT_AGENTS = [
+  "claude/sonnet-4.5",
+  "claude/opus-4.1",
+  "codex/gpt-5-codex-high",
+];
+const KNOWN_AGENT_NAMES = new Set(AGENT_CONFIGS.map((agent) => agent.name));
+const DEFAULT_AGENT_SELECTION = DEFAULT_AGENTS.filter((agent) =>
+  KNOWN_AGENT_NAMES.has(agent)
+);
+
+const AGENT_SELECTION_SCHEMA = z.array(z.string());
+
+const filterKnownAgents = (agents: string[]): string[] =>
+  agents.filter((agent) => KNOWN_AGENT_NAMES.has(agent));
+
+const parseStoredAgentSelection = (stored: string | null): string[] => {
+  if (!stored) {
+    return [];
+  }
+
+  try {
+    const parsed = JSON.parse(stored);
+    const result = AGENT_SELECTION_SCHEMA.safeParse(parsed);
+    if (!result.success) {
+      console.warn("Invalid stored agent selection", result.error);
+      return [];
+    }
+
+    return filterKnownAgents(result.data);
+  } catch (error) {
+    console.warn("Failed to parse stored agent selection", error);
+    return [];
+  }
+};
+
+export type UseTaskCreationFormOptions = {
+  teamSlugOrId: string;
+  searchEnvironmentId?: string;
+  enableGlobalKeydown?: boolean;
+};
+
+export type TaskCreationFormState = {
+  editorApiRef: React.RefObject<EditorApi | null>;
+  onTaskDescriptionChange: (value: string) => void;
+  onSubmit: () => void;
+  lexicalRepoUrl?: string;
+  lexicalEnvironmentId?: Id<"environments">;
+  lexicalBranch?: string;
+  projectOptions: SelectOption[];
+  selectedProject: string[];
+  onProjectChange: (projects: string[]) => void;
+  onProjectSearchPaste?: (value: string) => boolean | Promise<boolean>;
+  branchOptions: string[];
+  selectedBranch: string[];
+  onBranchChange: (branches: string[]) => void;
+  selectedAgents: string[];
+  onAgentChange: (agents: string[]) => void;
+  isCloudMode: boolean;
+  onCloudModeToggle: () => void;
+  isLoadingProjects: boolean;
+  isLoadingBranches: boolean;
+  providerStatus: ProviderStatusResponse | null;
+  canSubmit: boolean;
+  startTask: () => Promise<boolean>;
+  isEnvSelected: boolean;
+  selectedRepoFullName: string | null;
+  shouldShowWorkspaceSetup: boolean;
+};
+
+export function useTaskCreationForm({
+  teamSlugOrId,
+  searchEnvironmentId,
+  enableGlobalKeydown = true,
+}: UseTaskCreationFormOptions): TaskCreationFormState {
+  const { socket } = useSocket();
+  const { theme } = useTheme();
+  const { addTaskToExpand } = useExpandTasks();
+
+  const [selectedProject, setSelectedProject] = useState<string[]>(() => {
+    const stored = localStorage.getItem(`selectedProject-${teamSlugOrId}`);
+    return stored ? JSON.parse(stored) : [];
+  });
+  const [selectedBranch, setSelectedBranch] = useState<string[]>([]);
+
+  const [selectedAgents, setSelectedAgentsState] = useState<string[]>(() => {
+    const storedAgents = parseStoredAgentSelection(
+      localStorage.getItem("selectedAgents")
+    );
+
+    if (storedAgents.length > 0) {
+      return storedAgents;
+    }
+
+    return DEFAULT_AGENT_SELECTION.length > 0
+      ? [...DEFAULT_AGENT_SELECTION]
+      : [];
+  });
+  const selectedAgentsRef = useRef<string[]>(selectedAgents);
+
+  const setSelectedAgents = useCallback(
+    (agents: string[]) => {
+      selectedAgentsRef.current = agents;
+      setSelectedAgentsState(agents);
+    },
+    [setSelectedAgentsState]
+  );
+
+  const [taskDescription, setTaskDescription] = useState<string>("");
+  const [isCloudMode, setIsCloudMode] = useState<boolean>(() => {
+    const stored = localStorage.getItem("isCloudMode");
+    return stored ? JSON.parse(stored) : true;
+  });
+
+  const [, setDockerReady] = useState<boolean | null>(null);
+  const [providerStatus, setProviderStatus] =
+    useState<ProviderStatusResponse | null>(null);
+
+  const editorApiRef = useRef<EditorApi | null>(null);
+
+  const persistAgentSelection = useCallback((agents: string[]) => {
+    try {
+      const isDefaultSelection =
+        DEFAULT_AGENT_SELECTION.length > 0 &&
+        agents.length === DEFAULT_AGENT_SELECTION.length &&
+        agents.every(
+          (agent, index) => agent === DEFAULT_AGENT_SELECTION[index]
+        );
+
+      if (agents.length === 0 || isDefaultSelection) {
+        localStorage.removeItem("selectedAgents");
+      } else {
+        localStorage.setItem("selectedAgents", JSON.stringify(agents));
+      }
+    } catch (error) {
+      console.warn("Failed to persist agent selection", error);
+    }
+  }, []);
+
+  useEffect(() => {
+    if (searchEnvironmentId) {
+      const val = `env:${searchEnvironmentId}`;
+      setSelectedProject([val]);
+      localStorage.setItem(
+        `selectedProject-${teamSlugOrId}`,
+        JSON.stringify([val])
+      );
+      setIsCloudMode(true);
+      localStorage.setItem("isCloudMode", JSON.stringify(true));
+    }
+  }, [searchEnvironmentId, teamSlugOrId]);
+
+  const handleTaskDescriptionChange = useCallback((value: string) => {
+    setTaskDescription(value);
+  }, []);
+
+  const isEnvSelected = useMemo(
+    () => (selectedProject[0] || "").startsWith("env:"),
+    [selectedProject]
+  );
+
+  const branchesQuery = useQuery({
+    ...branchesQueryOptions({
+      teamSlugOrId,
+      repoFullName: selectedProject[0] || "",
+    }),
+    enabled: !!selectedProject[0] && !isEnvSelected,
+  });
+  const branchSummary = useMemo(() => {
+    const data = branchesQuery.data;
+    if (!data?.branches) {
+      return {
+        names: [] as string[],
+        defaultName: undefined as string | undefined,
+      };
+    }
+    const names = data.branches.map((branch) => branch.name);
+    const fromResponse = data.defaultBranch?.trim();
+    const flaggedDefault = data.branches.find((branch) => branch.isDefault)
+      ?.name;
+    const normalizedFromResponse =
+      fromResponse && names.includes(fromResponse) ? fromResponse : undefined;
+    const normalizedFlagged =
+      flaggedDefault && names.includes(flaggedDefault)
+        ? flaggedDefault
+        : undefined;
+
+    return {
+      names,
+      defaultName: normalizedFromResponse ?? normalizedFlagged,
+    };
+  }, [branchesQuery.data]);
+
+  const branchNames = branchSummary.names;
+  const remoteDefaultBranch = branchSummary.defaultName;
+
+  const handleProjectChange = useCallback(
+    (newProjects: string[]) => {
+      setSelectedProject(newProjects);
+      localStorage.setItem(
+        `selectedProject-${teamSlugOrId}`,
+        JSON.stringify(newProjects)
+      );
+      if (newProjects[0] !== selectedProject[0]) {
+        setSelectedBranch([]);
+      }
+      if ((newProjects[0] || "").startsWith("env:")) {
+        setIsCloudMode(true);
+        localStorage.setItem("isCloudMode", JSON.stringify(true));
+      }
+    },
+    [selectedProject, teamSlugOrId]
+  );
+
+  const handleBranchChange = useCallback((newBranches: string[]) => {
+    setSelectedBranch(newBranches);
+  }, []);
+
+  const handleAgentChange = useCallback(
+    (newAgents: string[]) => {
+      const normalizedAgents = filterKnownAgents(newAgents);
+      setSelectedAgents(normalizedAgents);
+      persistAgentSelection(normalizedAgents);
+    },
+    [persistAgentSelection, setSelectedAgents]
+  );
+
+  const reposByOrgQuery = useQuery({
+    ...convexQuery(api.github.getReposByOrg, { teamSlugOrId }),
+    refetchOnMount: "always",
+    refetchOnWindowFocus: false,
+  });
+  const reposByOrg = useMemo(
+    () => reposByOrgQuery.data || {},
+    [reposByOrgQuery.data]
+  );
+
+  const checkProviderStatus = useCallback(() => {
+    if (!socket) return;
+
+    socket.emit("check-provider-status", (response) => {
+      if (!response) return;
+      setProviderStatus(response);
+
+      if (response.success) {
+        const isRunning = response.dockerStatus?.isRunning;
+        if (typeof isRunning === "boolean") {
+          setDockerReady(isRunning);
+        }
+      }
+
+      const currentAgents = selectedAgentsRef.current;
+      if (currentAgents.length === 0) {
+        return;
+      }
+
+      const providers = response.providers;
+      if (!providers || providers.length === 0) {
+        const normalizedOnly = filterKnownAgents(currentAgents);
+        if (normalizedOnly.length !== currentAgents.length) {
+          setSelectedAgents(normalizedOnly);
+          persistAgentSelection(normalizedOnly);
+        }
+        return;
+      }
+
+      const availableAgents = new Set(
+        providers
+          .filter((provider) => provider.isAvailable)
+          .map((provider) => provider.name)
+      );
+
+      const normalizedAgents = filterKnownAgents(currentAgents);
+      const removedUnavailable = normalizedAgents.filter(
+        (agent) => !availableAgents.has(agent)
+      );
+
+      if (removedUnavailable.length === 0) {
+        return;
+      }
+
+      const filteredAgents = normalizedAgents.filter((agent) =>
+        availableAgents.has(agent)
+      );
+
+      setSelectedAgents(filteredAgents);
+      persistAgentSelection(filteredAgents);
+
+      if (removedUnavailable.length > 0) {
+        const uniqueMissing = Array.from(new Set(removedUnavailable));
+        if (uniqueMissing.length > 0) {
+          const label = uniqueMissing.length === 1 ? "model" : "models";
+          const verb = uniqueMissing.length === 1 ? "is" : "are";
+          toast.warning(
+            `${uniqueMissing.join(", ")} ${verb} not configured and was removed from the selection. Update credentials in Settings to use this ${label}.`
+          );
+        }
+      }
+    });
+  }, [persistAgentSelection, setSelectedAgents, socket]);
+
+  const createTask = useMutation(api.tasks.create).withOptimisticUpdate(
+    (localStore, args) => {
+      const currentTasks = localStore.getQuery(api.tasks.get, {
+        teamSlugOrId,
+      });
+
+      if (currentTasks !== undefined) {
+        const now = Date.now();
+        const optimisticTask = {
+          _id: createFakeConvexId() as Doc<"tasks">["_id"],
+          _creationTime: now,
+          text: args.text,
+          description: args.description,
+          projectFullName: args.projectFullName,
+          baseBranch: args.baseBranch,
+          worktreePath: args.worktreePath,
+          isCompleted: false,
+          isArchived: false,
+          createdAt: now,
+          updatedAt: now,
+          images: args.images,
+          userId: "optimistic",
+          teamId: teamSlugOrId,
+          environmentId: args.environmentId,
+        };
+
+        localStore.setQuery(api.tasks.get, { teamSlugOrId }, [
+          optimisticTask,
+          ...currentTasks,
+        ]);
+      }
+    }
+  );
+  const generateUploadUrl = useMutation(api.storage.generateUploadUrl);
+  const addManualRepo = useAction(api.github_http.addManualRepo);
+
+  const effectiveSelectedBranch = useMemo(() => {
+    if (selectedBranch.length > 0) {
+      return selectedBranch;
+    }
+    if (branchNames.length === 0) {
+      return [];
+    }
+    const fallbackBranch = branchNames.includes("main")
+      ? "main"
+      : branchNames.includes("master")
+        ? "master"
+        : branchNames[0];
+    const preferredBranch =
+      remoteDefaultBranch && branchNames.includes(remoteDefaultBranch)
+        ? remoteDefaultBranch
+        : fallbackBranch;
+    return [preferredBranch];
+  }, [selectedBranch, branchNames, remoteDefaultBranch]);
+
+  const handleStartTask = useCallback(async (): Promise<boolean> => {
+    if (!isEnvSelected && !isCloudMode) {
+      if (socket) {
+        const ready = await new Promise<boolean>((resolve) => {
+          socket.emit("check-provider-status", (response) => {
+            const isRunning = !!response?.dockerStatus?.isRunning;
+            if (typeof isRunning === "boolean") {
+              setDockerReady(isRunning);
+            }
+            resolve(isRunning);
+          });
+        });
+
+        if (!ready) {
+          toast.error("Docker is not running. Start Docker Desktop.");
+          return false;
+        }
+      } else {
+        console.error("Cannot verify Docker status: socket not connected");
+        toast.error(
+          "Cannot verify Docker status. Please ensure the server is running."
+        );
+        return false;
+      }
+    }
+
+    if (!selectedProject[0] || !taskDescription.trim()) {
+      console.error("Please select a project and enter a task description");
+      return false;
+    }
+    if (!socket) {
+      console.error("Socket not connected");
+      return false;
+    }
+
+    const branch = effectiveSelectedBranch[0];
+    const projectFullName = selectedProject[0];
+    const envSelected = projectFullName.startsWith("env:");
+    const environmentId = envSelected
+      ? (projectFullName.replace(/^env:/, "") as Id<"environments">)
+      : undefined;
+
+    try {
+      const content = editorApiRef.current?.getContent();
+      const images = content?.images || [];
+
+      const uploadedImages = await Promise.all(
+        images.map(
+          async (image: {
+            src: string;
+            fileName?: string;
+            altText: string;
+          }) => {
+            const base64Data = image.src.split(",")[1] || image.src;
+            const byteCharacters = atob(base64Data);
+            const byteNumbers = new Array(byteCharacters.length);
+            for (let i = 0; i < byteCharacters.length; i++) {
+              byteNumbers[i] = byteCharacters.charCodeAt(i);
+            }
+            const byteArray = new Uint8Array(byteNumbers);
+            const blob = new Blob([byteArray], { type: "image/png" });
+            const uploadUrl = await generateUploadUrl({
+              teamSlugOrId,
+            });
+            const result = await fetch(uploadUrl, {
+              method: "POST",
+              headers: { "Content-Type": blob.type },
+              body: blob,
+            });
+            const { storageId } = await result.json();
+
+            return {
+              storageId,
+              fileName: image.fileName,
+              altText: image.altText,
+            };
+          }
+        )
+      );
+
+      setTaskDescription("");
+      handleTaskDescriptionChange("");
+      if (editorApiRef.current?.clear) {
+        editorApiRef.current.clear();
+      }
+
+      const taskId = await createTask({
+        teamSlugOrId,
+        text: content?.text || taskDescription,
+        projectFullName: envSelected ? undefined : projectFullName,
+        baseBranch: envSelected ? undefined : branch,
+        images: uploadedImages.length > 0 ? uploadedImages : undefined,
+        environmentId,
+      });
+
+      addTaskToExpand(taskId);
+
+      const repoUrl = envSelected
+        ? undefined
+        : `https://github.com/${projectFullName}.git`;
+
+      const handleStartTaskAck = (
+        response: TaskAcknowledged | TaskStarted | TaskError
+      ) => {
+        if ("error" in response) {
+          console.error("Task start error:", response.error);
+          toast.error(`Task start error: ${JSON.stringify(response.error)}`);
+          return;
+        }
+
+        attachTaskLifecycleListeners(socket, response.taskId, {
+          onStarted: (payload) => {
+            console.log("Task started:", payload);
+          },
+          onFailed: (payload) => {
+            toast.error(`Task failed to start: ${payload.error}`);
+          },
+        });
+        console.log("Task acknowledged:", response);
+      };
+
+      socket.emit(
+        "start-task",
+        {
+          ...(repoUrl ? { repoUrl } : {}),
+          ...(envSelected ? {} : { branch }),
+          taskDescription: content?.text || taskDescription,
+          projectFullName,
+          taskId,
+          selectedAgents:
+            selectedAgents.length > 0 ? selectedAgents : undefined,
+          isCloudMode: envSelected ? true : isCloudMode,
+          ...(environmentId ? { environmentId } : {}),
+          images: images.length > 0 ? images : undefined,
+          theme,
+        },
+        handleStartTaskAck
+      );
+      console.log("Task created:", taskId);
+      return true;
+    } catch (error) {
+      console.error("Error starting task:", error);
+      return false;
+    }
+  }, [
+    isEnvSelected,
+    isCloudMode,
+    socket,
+    selectedProject,
+    taskDescription,
+    effectiveSelectedBranch,
+    generateUploadUrl,
+    teamSlugOrId,
+    createTask,
+    addTaskToExpand,
+    selectedAgents,
+    theme,
+    handleTaskDescriptionChange,
+  ]);
+
+  useEffect(() => {
+    checkProviderStatus();
+
+    const interval = setInterval(() => {
+      checkProviderStatus();
+    }, 5000);
+
+    const handleFocus = () => checkProviderStatus();
+    window.addEventListener("focus", handleFocus);
+
+    return () => {
+      clearInterval(interval);
+      window.removeEventListener("focus", handleFocus);
+    };
+  }, [checkProviderStatus]);
+
+  const environmentsQuery = useQuery(
+    convexQuery(api.environments.list, { teamSlugOrId })
+  );
+
+  const projectOptions = useMemo(() => {
+    const repoDocs = Object.values(reposByOrg || {}).flatMap((repos) => repos);
+    const uniqueRepos = repoDocs.reduce((acc, repo) => {
+      const existing = acc.get(repo.fullName);
+      if (!existing) {
+        acc.set(repo.fullName, repo);
+        return acc;
+      }
+      const existingActivity =
+        existing.lastPushedAt ?? Number.NEGATIVE_INFINITY;
+      const candidateActivity = repo.lastPushedAt ?? Number.NEGATIVE_INFINITY;
+      if (candidateActivity > existingActivity) {
+        acc.set(repo.fullName, repo);
+      }
+      return acc;
+    }, new Map<string, Doc<"repos">>());
+    const sortedRepos = Array.from(uniqueRepos.values()).sort((a, b) => {
+      const aPushedAt = a.lastPushedAt ?? Number.NEGATIVE_INFINITY;
+      const bPushedAt = b.lastPushedAt ?? Number.NEGATIVE_INFINITY;
+      if (aPushedAt !== bPushedAt) {
+        return bPushedAt - aPushedAt;
+      }
+      return a.fullName.localeCompare(b.fullName);
+    });
+    const repoOptions = sortedRepos.map((repo) => ({
+      label: repo.fullName,
+      value: repo.fullName,
+      icon: (
+        <GitHubIcon className="w-4 h-4 text-neutral-600 dark:text-neutral-300" />
+      ),
+      iconKey: "github",
+    }));
+
+    const envOptions = (environmentsQuery.data || []).map((env) => ({
+      label: `${env.name}`,
+      value: `env:${env._id}`,
+      icon: (
+        <Tooltip>
+          <TooltipTrigger asChild>
+            <span>
+              <ServerIcon className="w-4 h-4 text-neutral-600 dark:text-neutral-300" />
+            </span>
+          </TooltipTrigger>
+          <TooltipContent>Environment: {env.name}</TooltipContent>
+        </Tooltip>
+      ),
+      iconKey: "environment",
+    }));
+
+    const options: SelectOption[] = [];
+    if (envOptions.length > 0) {
+      options.push({
+        label: "Environments",
+        value: "__heading-env",
+        heading: true,
+      });
+      options.push(...envOptions);
+    }
+    if (repoOptions.length > 0) {
+      options.push({
+        label: "Repositories",
+        value: "__heading-repo",
+        heading: true,
+      });
+      options.push(...repoOptions);
+    }
+
+    return options;
+  }, [reposByOrg, environmentsQuery.data]);
+
+  const selectedRepoFullName = useMemo(() => {
+    if (!selectedProject[0] || isEnvSelected) return null;
+    return selectedProject[0];
+  }, [selectedProject, isEnvSelected]);
+
+  const shouldShowWorkspaceSetup = !!selectedRepoFullName && !isEnvSelected;
+
+  const handleProjectSearchPaste = useCallback(
+    async (input: string) => {
+      try {
+        const result = await addManualRepo({
+          teamSlugOrId,
+          repoUrl: input,
+        });
+
+        if (result.success) {
+          await reposByOrgQuery.refetch();
+
+          setSelectedProject([result.fullName]);
+          localStorage.setItem(
+            `selectedProject-${teamSlugOrId}`,
+            JSON.stringify([result.fullName])
+          );
+
+          toast.success(`Added ${result.fullName} to repositories`);
+          return true;
+        }
+
+        return false;
+      } catch (error) {
+        if (
+          error instanceof Error &&
+          error.message &&
+          !error.message.includes("Invalid GitHub")
+        ) {
+          toast.error(error.message);
+        }
+        return false;
+      }
+    },
+    [addManualRepo, teamSlugOrId, reposByOrgQuery]
+  );
+
+  useEffect(() => {
+    if (!socket) return;
+
+    const handleVSCodeSpawned = (data: {
+      instanceId: string;
+      url: string;
+      workspaceUrl: string;
+      provider: string;
+    }) => {
+      console.log("VSCode spawned:", data);
+    };
+
+    socket.on("vscode-spawned", handleVSCodeSpawned);
+
+    return () => {
+      socket.off("vscode-spawned", handleVSCodeSpawned);
+    };
+  }, [socket]);
+
+  useEffect(() => {
+    if (!socket) return;
+
+    const handleDefaultRepo = (data: {
+      repoFullName: string;
+      branch?: string;
+      localPath: string;
+    }) => {
+      setSelectedProject([data.repoFullName]);
+      localStorage.setItem(
+        `selectedProject-${teamSlugOrId}`,
+        JSON.stringify([data.repoFullName])
+      );
+
+      if (data.branch) {
+        setSelectedBranch([data.branch]);
+      }
+    };
+
+    socket.on("default-repo", handleDefaultRepo);
+
+    return () => {
+      socket.off("default-repo", handleDefaultRepo);
+    };
+  }, [socket, teamSlugOrId]);
+
+  useEffect(() => {
+    if (!enableGlobalKeydown) return;
+
+    const handleGlobalKeyDown = (e: KeyboardEvent) => {
+      const activeElement = document.activeElement;
+      const isEditor =
+        activeElement?.getAttribute("data-cmux-input") === "true";
+      const isCommentInput = activeElement?.id === "cmux-comments-root";
+      if (
+        !isEditor &&
+        (activeElement?.tagName === "INPUT" ||
+          activeElement?.tagName === "TEXTAREA" ||
+          activeElement?.getAttribute("contenteditable") === "true" ||
+          activeElement?.closest('[contenteditable="true"]') ||
+          isCommentInput)
+      ) {
+        return;
+      }
+
+      if (
+        e.ctrlKey ||
+        e.metaKey ||
+        e.altKey ||
+        e.key.length !== 1 ||
+        e.key === " " ||
+        e.key === "Escape" ||
+        e.key === "Enter" ||
+        e.key === "Tab"
+      ) {
+        return;
+      }
+
+      if (editorApiRef.current?.focus) {
+        editorApiRef.current.focus();
+
+        if (editorApiRef.current.insertText) {
+          editorApiRef.current.insertText(e.key);
+        }
+      }
+    };
+
+    document.addEventListener("keydown", handleGlobalKeyDown);
+
+    return () => {
+      document.removeEventListener("keydown", handleGlobalKeyDown);
+    };
+  }, [enableGlobalKeydown]);
+
+  const handleSubmit = useCallback(() => {
+    if (selectedProject[0] && taskDescription.trim()) {
+      void handleStartTask();
+    }
+  }, [selectedProject, taskDescription, handleStartTask]);
+
+  const handleCloudModeToggle = useCallback(() => {
+    if (isEnvSelected) return;
+    const newMode = !isCloudMode;
+    setIsCloudMode(newMode);
+    localStorage.setItem("isCloudMode", JSON.stringify(newMode));
+  }, [isCloudMode, isEnvSelected]);
+
+  const lexicalEnvironmentId = useMemo(() => {
+    if (!selectedProject[0] || !isEnvSelected) return undefined;
+    return selectedProject[0].replace(/^env:/, "") as Id<"environments">;
+  }, [selectedProject, isEnvSelected]);
+
+  const lexicalRepoUrl = useMemo(() => {
+    if (!selectedProject[0]) return undefined;
+    if (isEnvSelected) return undefined;
+    return `https://github.com/${selectedProject[0]}.git`;
+  }, [selectedProject, isEnvSelected]);
+
+  const lexicalBranch = useMemo(
+    () => effectiveSelectedBranch[0],
+    [effectiveSelectedBranch]
+  );
+
+  const canSubmit = useMemo(() => {
+    if (!selectedProject[0]) return false;
+    if (!taskDescription.trim()) return false;
+    if (selectedAgents.length === 0) return false;
+    if (isEnvSelected) return true;
+    return !!effectiveSelectedBranch[0];
+  }, [
+    selectedProject,
+    taskDescription,
+    selectedAgents,
+    isEnvSelected,
+    effectiveSelectedBranch,
+  ]);
+
+  return {
+    editorApiRef,
+    onTaskDescriptionChange: handleTaskDescriptionChange,
+    onSubmit: handleSubmit,
+    lexicalRepoUrl,
+    lexicalEnvironmentId,
+    lexicalBranch,
+    projectOptions,
+    selectedProject,
+    onProjectChange: handleProjectChange,
+    onProjectSearchPaste: handleProjectSearchPaste,
+    branchOptions: branchNames,
+    selectedBranch: effectiveSelectedBranch,
+    onBranchChange: handleBranchChange,
+    selectedAgents,
+    onAgentChange: handleAgentChange,
+    isCloudMode,
+    onCloudModeToggle: handleCloudModeToggle,
+    isLoadingProjects: reposByOrgQuery.isLoading,
+    isLoadingBranches: branchesQuery.isPending,
+    providerStatus,
+    canSubmit,
+    startTask: handleStartTask,
+    isEnvSelected,
+    selectedRepoFullName,
+    shouldShowWorkspaceSetup,
+  };
+}
